[Refactor] Move descriptors_ext into RuntimeCore
