+++ be/module_boundary_manifest.json
@@ -246,6 +246,7 @@
         "be/src/runtime/current_thread.h",
         "be/src/runtime/descriptor_helper.h",
         "be/src/runtime/descriptors.h",
+        "be/src/runtime/descriptors_ext.h",
         "be/src/runtime/descriptors_fwd.h",
         "be/src/runtime/exception.h",
         "be/src/runtime/exec_env_fwd.h",
