+++ be/src/base/CMakeLists.txt
@@ -35,6 +35,7 @@ ADD_BE_LIB(Base
   debug/trace.cpp
   debug/trace_metrics.cpp
   coding.cpp
+  base64.cpp
   gc/gc_helper.cpp
   memory/jemalloc_allocator.cpp
   memory/malloc_allocator.cpp
