+++ be/src/exprs/bitmap_functions.cpp
@@ -14,6 +14,7 @@
 
 #include "exprs/bitmap_functions.h"
 
+#include "base/base64.h"
 #include "base/phmap/phmap.h"
 #include "base/string/string_parser.hpp"
 #include "column/array_column.h"
@@ -22,7 +23,6 @@
 #include "column/column_viewer.h"
 #include "column/nullable_column.h"
 #include "common/config_expr_fwd.h"
-#include "exprs/base64.h"
 #include "exprs/binary_function.h"
 #include "exprs/function_context.h"
 #include "exprs/unary_function.h"
