+++ be/src/service/service_be/arrow_flight_sql_service.cpp
@@ -20,11 +20,11 @@
 #include <base/utility/arrow_utils.h>
 #include <exec/pipeline/query_context.h>
 
+#include "base/base64.h"
 #include "base/uid_util.h"
 #include "common/status.h"
 #include "common/system/backend_options.h"
 #include "exec/arrow_flight_batch_reader.h"
-#include "exprs/base64.h"
 #include "runtime/exec_env.h"
 
 namespace starrocks {
