+++ be/src/udf/python/callstub.cpp
@@ -21,12 +21,12 @@
 #include "arrow/buffer.h"
 #include "arrow/flight/client.h"
 #include "arrow/type.h"
+#include "base/base64.h"
 #include "base/utility/arrow_utils.h"
 #include "base/utility/defer_op.h"
 #include "common/config_udf_fwd.h"
 #include "common/status.h"
 #include "common/statusor.h"
-#include "exprs/base64.h"
 #include "rapidjson/document.h"
 #include "rapidjson/rapidjson.h"
 #include "rapidjson/stringbuffer.h"
